--- app/src/components/elements/NavBar/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useNavRoutes';
