--- app/src/components/elements/SearchBar/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useSearchBar';
