--- app/src/pages/HomePage/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useHomePageDashboard';
