+++ app/src/pages/HomePage/index.tsx
@@ -2,7 +2,7 @@ import { Seo } from '@components/elements/Seo';
 import { Dashboard } from '@components/templates/Dashboard';
 import { withFooter } from '@hoc/withFooter';
 import { withHead } from '@hoc/withHead';
-import { useHomePageDashboard } from './hooks';
+import { useHomePageDashboard } from './hooks/useHomePageDashboard';
 
 const HomePage = () => {
   return <Dashboard {...useHomePageDashboard()} />;
