--- app/src/pages/PageSkeletons/HomePageSkeleton/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useHomePageDashboardSkeleton';
