+++ app/src/pages/PageSkeletons/HomePageSkeleton/index.tsx
@@ -1,5 +1,5 @@
 import { DashboardSkeleton } from '@components/templates/DashboardSkeleton';
-import { useHomePageDashboardSkeleton } from './hooks';
+import { useHomePageDashboardSkeleton } from './hooks/useHomePageDashboardSkeleton';
 
 export const HomePageSkeleton = () => {
   return <DashboardSkeleton {...useHomePageDashboardSkeleton()} />;
