--- app/src/pages/PageSkeletons/ProfileEvalTabSkeleton/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useProfileEvalTabDashboardSkeleton';
