+++ app/src/pages/PageSkeletons/ProfileEvalTabSkeleton/index.tsx
@@ -1,5 +1,5 @@
 import { DashboardSkeleton } from '@components/templates/DashboardSkeleton';
-import { useProfileEvalTabDashboardSkeleton } from './hooks';
+import { useProfileEvalTabDashboardSkeleton } from './hooks/useProfileEvalTabDashboardSkeleton';
 
 export const ProfileEvalTabSkeleton = () => {
   return <DashboardSkeleton {...useProfileEvalTabDashboardSkeleton()} />;
