--- app/src/pages/PageSkeletons/ProfilePageSkeleton/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useProfileGeneralTabDashboardSkeleton';
-export * from './useProfilePageDashboardSkeleton';
