+++ app/src/pages/PageSkeletons/ProfilePageSkeleton/index.tsx
@@ -1,9 +1,7 @@
 import { VStack } from '@components/common';
 import { DashboardSkeleton } from '@components/templates/DashboardSkeleton';
-import {
-  useProfileGeneralTabDashboardSkeleton,
-  useProfilePageDashboardSkeleton,
-} from './hooks';
+import { useProfileGeneralTabDashboardSkeleton } from './hooks/useProfileGeneralTabDashboardSkeleton';
+import { useProfilePageDashboardSkeleton } from './hooks/useProfilePageDashboardSkeleton';
 
 export const ProfilePageSkeleton = () => {
   return (
