--- app/src/pages/ProfileEvalTab/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useProfileEvalTabDashboard';
