+++ app/src/pages/ProfileEvalTab/index.tsx
@@ -1,5 +1,5 @@
 import { Dashboard } from '@components/templates/Dashboard';
-import { useProfileEvalTabDashboard } from './hooks';
+import { useProfileEvalTabDashboard } from './hooks/useProfileEvalTabDashboard';
 
 const ProfileEvalTab = () => {
   return <Dashboard {...useProfileEvalTabDashboard()} />;
