--- app/src/pages/ProfileGeneralTab/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useProfileGeneralTabDashboard';
