+++ app/src/pages/ProfileGeneralTab/index.tsx
@@ -1,5 +1,5 @@
 import { Dashboard } from '@components/templates/Dashboard';
-import { useProfileGeneralTabDashboard } from './hooks';
+import { useProfileGeneralTabDashboard } from './hooks/useProfileGeneralTabDashboard';
 
 const ProfileGeneralTab = () => {
   return <Dashboard {...useProfileGeneralTabDashboard()} />;
